+++ libNOM.io/Delegates/NotifyWriteEventHandler.cs
@@ -1,4 +1,0 @@
-﻿namespace libNOM.io.Delegates;
-
-
-public delegate void NotifyWriteEventHandler();