+++ libNOM.io/Platform_FileOperation_Transfer.cs
@@ -167,9 +167,12 @@ public abstract partial class Platform : IPlatform, IEquatable<Platform>
 
                 TransferOwnership(Destination, sourceTransferData);
 
-                // This "if" is not really useful in this method but properly implemented nonetheless.
+                // Write() or/and Invoke().
                 if (write)
                     Write(Destination, Source.LastWriteTime ?? DateTimeOffset.Now);
+                else
+                    Destination.PropertiesChangedCallback.Invoke();
+                Destination.JsonChangedCallback.Invoke();
             }
     }